+++ Makefile.config
@@ -86,7 +86,7 @@ BOARD          ?= ciaa_sim_ia32
 # examples/adc_dac                   (example using adc and dac)
 # examples/blinking                  (simple example with rtos and posix)
 # examples/blinking_echo             (example with rtos and posix)
-# examples/blinking_modbus           (example with rtos, posix and using modbus)
+# examples/blinking_modbus_slave     (example with rtos, posix and using modbus slave)
 # examples/blinking_modbus_master    (example with rtos, posix and using modbus master)
 # examples/blinking_lwip             (blinking and tcp echo using lwip)
 # examples/rtos_example              (example using rtos)
